Add airflowignore file (Recidiviz/recidiviz-data#31152)
## Description of the change

Only look for DAG definitions in top level files to cut down on parse
time + improve scheduler performance.
When I added this to my test environment it significantly cut down my
parse time but there was also a lot of junk in my test env
![Screenshot 2024-06-24 at 10 45
51 AM](https://github.com/Recidiviz/recidiviz-data/assets/149205316/3981439e-b16f-475f-9e45-fbc249b59fcb)
"""
When searching for DAGs inside the DAG_FOLDER, Airflow only considers
Python files that contain the strings airflow and dag
(case-insensitively) as an optimization.

You can also provide an .airflowignore file inside your DAG_FOLDER, or
any of its subfolders, which describes patterns of files for the loader
to ignore. It covers the directory it’s in plus all subfolders
underneath it.
"""

https://airflow.apache.org/docs/apache-airflow/stable/core-concepts/dags.html

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d3cf24250eb3279a3d3b3752e153fcc463c96e91