[Justice Counts] Automatically Synchronize User Access Between Super Agency and Child Agencies (Recidiviz/recidiviz-data#31147)
## Description of the change

This update ensures that when users are added to or removed from a super
agency, their access is automatically synchronized with all child
agencies. This means that any user added to the super agency will also
be granted access to the child agencies with the same role. Conversely,
removing a user from the super agency will also revoke their access from
the child agencies. This functionality existed in the legacy admin
panel, but I forgot to migrate it over in the new admin panel. This also
already happens when a user adds a new user to their super agency via
agency settings.

https://github.com/Recidiviz/recidiviz-data/assets/19961693/9616cb93-a8a1-4336-9a30-d840bbe15266

This works with the accompanied [FE
PR](https://github.com/Recidiviz/justice-counts/pull/1421).

## Related issues

Closes Recidiviz/recidiviz-data#30561

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 85fe19cad8a9e01832deb0e563ff9045c19e9f54