[Normalization] More typing cleanup to prep for v2 normalization entities (Recidiviz/recidiviz-data#31010)
## Description of the change

When we migrate over to the entities defined in
normalized_entities_v2.py, the normalized version of each entity will no
longer be a subclass of it's corresponding pre-normalization entity.
This PR introduces more mypy changes to places where we were referencing
the non-normalized version of the entity but can/should be referencing
the normalized version or updates typing to indicate we could reference
either.

Most notably, this updates supervision period start_date to be
non-optional for pre-normalization entities (checked in BQ and it's
always non-null).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#30075

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 41060e0d7ae3173652389ee504c4039e00960e8d