[US_PA] Pull status into supervision periods view so that we can map IN_CUSTODY supervision level (Recidiviz/recidiviz-data#31251)
## Description of the change

Pull the status_code variable into supervision periods view results so
that we can map IN_CUSTODY and ABSCONSION supervision levels.

This revises the _v4 version of the view, which is what lives in prod
(_v5 is the refactored version currently gated to staging).

Testing:
- [x] Updated unit tests
- [x] Confirmed that the values in the other columns of the ingest view
results don't change at all
- [x] Run mappings succeeds
- [x] Run ingest pipeline (in progress)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b4bb6d92f105bed98aa792e5b3f86e3098a42873