[Dataflow] Remove reliance on schema.py / entity subclass relationships in extractor utils (part 1) (Recidiviz/recidiviz-data#31475)
## Description of the change

The v2 normalized entities will not be subclasses of the entities in
state/entities.py. There are a number of places in extractor_utils.py
where we're relying on the subclass relationship between the two sets of
entities and we will need to refactor to remove that dependency. This PR
does some initial simple refactors to simplify some of the extractor
code to eliminate usage of base/subclass distinction and just use table
ids and primary key column names directly (those do not change between
entity and normalized entity classes).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#30075

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d7fa8349013c333d51d172637cdd97ac16b56734