add new aggregated metrics views that separate out caseload categories/types (Recidiviz/recidiviz-data#31468)
## Description of the change

Adds new aggregated metrics views that contain the values for each
metric we use for insights separated out by caseload category/type.

Loaded to sandbox at `dana_insights_agg`.
`supervision_insights_caseload_category_aggregated_metrics_materialized`
is the one we'll use from the product views to determine the correct
benchmarks to show- the other 3 `supervision_insights_caseload_category`
ones aren't needed in the product views but are used to create the one
that we do use.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#31467

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e4c593775fd75029119075bac2f436af6fb5273d