TASK_COMPLETED event query (Recidiviz/recidiviz-data#31207)
## Description of the change

> Implemented new query to add "ineligible_criteria" to the json in
event_attributes in the person_events table. It joins a subquery that
handles the unnest of ineligible_criteria in all_tasks table and
aggregates all the strings that belong to the same observation. We join
that table, named a_t in the code, with
all_completion_events_materialized on completion_event_type, person_id,
and checking that the completion_event_date is between start_date and
end_date. Checked that persons_events has the same number of
observations after we add ineligible_criteria to event_attributes

## Type of change

Type: Feature

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: f312c88f994a46e007f755bfbbf2875649b801de