[Dataflow] Remove reliance on schema.py / entity subclass relationships in extractor utils (part 2) (Recidiviz/recidiviz-data#31533)
## Description of the change

Refactors the `ExtractRootEntityDataForPipeline` transform to support
the new v2 normalized entities that are not subclasses of the entities
defined in entities.py.

This refactor eliminates all reliance on the `schema.py` schema
(fundamentally tied to the non-normalized versions of the entities) by
refactoring to instead use the new `is_many_to_many_relationship` etc
utils.

This was tricky to do as a distinct change *before* we actually migrate
over to the v2 entities because right now, we sometimes are hydrating a
*MIX* of `NormalizedState*` and `State*` entities (😱). Whenever we're
trying to understand a relationship between two entity types, we need to
compare using the `entities.py` version. In order to make this easy, I
introduced a temporary `self._unified_schema_class` map that allows us
to map any entity class to the corresponding one in the `entities.py`
schema.

TODO before merging:
- [x] Run normalization pipelines to test that reading in data from the
`state` dataset isn't broken
- [x] Run metric pipelines to test that reading in data from the
`normalized_state` dataset isn't broken

Ran a full suite of pipelines and confirmed the outputs are identical /
differences can be explained by data that came in overnight.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#30075

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 23f0bfdb79c83fdb80548ee869ac4283e3fb642a