[Case Note Search] Add authentication validation to the case note search endpoint. (Recidiviz/recidiviz-data#30833)
## Description of the change

Add authentication validation to the case note search endpoint.

[Zenhub
Ticket](https://app.zenhub.com/workspaces/ai-explorations-6698285eed81380791facbd7/issues/gh/recidiviz/recidiviz-data/31514)

In this PR, we are also adding a python script that can be used for
querying the endpoint.

## Testing
Ran the app locally and queried it using the new prototypes.request_api
script. Verified that it requires a valid token to call the endpoint.

GitOrigin-RevId: c2e496e8c0340b56f40f0499591b319ea49777e7