[US_AZ] Refactor supervision periods using new data (Recidiviz/recidiviz-data#31488)
## Description of the change

This is an XXL refactor of the Arizona supervision periods ingest
process. The old query can be ignored - this should be treated like new
ingest.

The new query is based off of the 960 report in AZ, which is what ADCRR
uses to gauge how many people are on supervision, where they are
located, and to which officer's caseload they are assigned. The source
tables used in this new query are the same as the source tables used in
the query that generates that report. Examples of that report can be
found at go/Arizona -> Validation Data -> 960 Reports

I believe this ingest view structure may be generalizable to other
states and period entities, at least to some extent. The general
structure pulls in one critical date per CTE, then creates a large table
with all of those dates unioned together. The attributes assigned in one
source table are then carried forward where appropriate to other rows
(i.e. supervision levels are carried forward into rows where supervising
officers are assigned), and then periods are created using that master
table.

Each component of a full sandbox pipeline (ingest through metrics)
succeeded using the `full_sandbox_pipeline.sh` script.

Tests:

- [x] ingest view query is deterministic
- [x] external ID values are unique
- [x] run mappings script passes
- [x] sandbox ingest
[pipeline](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-07-17_08_26_17-10999984191883280695;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))
passes

Sanity checks:
- [x] < 6000 open periods
- [x] no open periods from > 10 years ago
- [x] no `PERSON_ID` is associated with more than one open period

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#30389

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 63235268b31a0d85f9e81bf6163214bce71e2848