[raw data import dag] `coalesce_results_and_errors` task (Recidiviz/recidiviz-data#31540)
## Description of the change

adds `coalesce_results_and_errors` task and tests. the idea behind this
task is to take in the reconcile the starting
`RawBigQueryFileMetadataSummary` with the results and errors to generate
--
- `import_session_summaries`: a list of `ImportSessionSummary` objects
to be persisted to the operations database
- `processed_paths_to_rename`: a list of `GcsfsFilePath` objects that
need their processed state renamed from unprocessed to processed after
successfully being imported
- `temporary_paths_to_clean`: a list of `GcsfsFilePath` objects that
need to be deleted
- `temporary_tables_to_clean`: a list of `BigQueryAddress` objects that
need to be deleted

each of the outputs will be handed down to a child task that will be
responsible for doing the actual task; (code for each task to come in
subsequent PRs). see Recidiviz/recidiviz-data#31651 to see how this task relates to its
downstream tasks

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#30169

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7c3a4f716eb04844d93810142832f3dca5c09e06